add in usb driver, minus all the clients (compiles, not enabled)
so, the pending service interrupt turns out to be architectural (kinda), and it looks as if its meant
to allow more software scheduling around prioritization (you would post such a thing in a high-priority
interrupt to defer the work). so, we'll ignore that until we decide how we want to scheudle this thing.